Update README.md
The SkillUp project is designed to offer a modern online learning environment that empowers individuals to enhance their knowledge and skills. Our goal is to create an effective framework that supports continuous education and growth, both personally and professionally. The platform features a wide range of courses, engaging learning activities, and an inclusive atmosphere to foster a productive and immersive learning experience. The solution is built using cutting-edge technologies such as React.js, Express.js, Node.js, and PostgreSQL to ensure a smooth and scalable user experience.